fix: moved esc listener to modal component.
